rename files accordingly
