+++ src/mindlessgen/qm/__init__.py
@@ -5,7 +5,7 @@ This module contains all QM-related functions and classes.
 from .base import QMMethod
 from .xtb import XTB, get_xtb_path
 from .orca import ORCA, get_orca_path
-from .gp3 import GXTB, get_gxtb_path
+from .gxtb import GXTB, get_gxtb_path
 
 __all__ = [
     "XTB",
